fuzz: add fuzzing of hid_linux, initially only enumeration
This makes use of /dev/uhid to create a dummy user-space HID transport
driver. This makes it possible for the fuzzer to create/destroy a number
of virtual devices which are then found by udev. Opening /dev/uhid for
writing requires root privileges. The privileges can be dropped as soon
as they are not needed any more.

To run this in a docker container requires not only the --privileged
flag, but also using a bind-mount for /dev because the virtual devices'
paths (/dev/hidraw*) does not seem to be passed through to the container
without it. Ideally, we would want to be a little bit more resticted
about what devices are passed through, but I have not found any suitable
options for that.